[REG-1778] Add Mouse clicker bot segment support
