+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/IBotActionData.cs
@@ -4,11 +4,19 @@ namespace RegressionGames.StateRecorder.BotSegments.Models
 {
     public interface IBotActionData
     {
+        /**
+         * Called at least once per frame
+         */
+        public void ProcessAction();
+
         public void WriteToStringBuilder(StringBuilder stringBuilder);
 
         public void ReplayReset();
 
-        public bool IsCompleted();
+        /**
+         * Returns null if the action runs until the keyframecriteria match
+         */
+        public bool? IsCompleted();
 
         public int EffectiveApiVersion();
     }
