+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/JsonConverters/RectIntJsonConverter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b0df7f9fa71d4b11a6725ff8ea69e4d8
+timeCreated: 1718203929
