+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/KeyboardEventSender.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 402d2ee500d346b1a48a7d27a8d98c1e
+timeCreated: 1718207480
