+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/KeyState.cs
@@ -1,0 +1,8 @@
+﻿namespace RegressionGames.StateRecorder.Models
+{
+    public enum KeyState
+    {
+        Up,
+        Down
+    }
+}
