+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/Models/KeyState.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4396a969ca8641da8d27a88ef9062f8d
+timeCreated: 1718206913
