+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/MouseEventSender.cs
@@ -1,0 +1,451 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using RegressionGames.RGLegacyInputUtility;
+using RegressionGames.StateRecorder.Models;
+using UnityEngine;
+using UnityEngine.InputSystem;
+using UnityEngine.InputSystem.Controls;
+using UnityEngine.InputSystem.LowLevel;
+using UnityEngine.InputSystem.Utilities;
+
+namespace RegressionGames.StateRecorder
+{
+    public static class MouseEventSender
+    {
+
+        private static Vector2? _lastMousePosition;
+
+        private static IDisposable _mouseEventHandler;
+
+        // ReSharper disable once InconsistentNaming
+        private static readonly RecordedGameObjectStatePathEqualityComparer _recordedGameObjectStatePathEqualityComparer = new();
+
+        public static void Reset()
+        {
+            _mouseEventHandler?.Dispose();
+            _mouseEventHandler = null;
+        }
+
+        public static InputDevice GetMouse(bool forceListener = false)
+        {
+            var mouse = InputSystem.devices.FirstOrDefault(a => a.name == "RGVirtualMouse");
+            var created = false;
+            if (mouse == null)
+            {
+                mouse = InputSystem.AddDevice<Mouse>("RGVirtualMouse");
+                created = true;
+            }
+
+            if (!mouse.enabled)
+            {
+                InputSystem.EnableDevice(mouse);
+            }
+
+            if (forceListener || created)
+            {
+                _mouseEventHandler = InputSystem.onEvent.ForDevice(mouse).Call(e =>
+                {
+                    var positionControl = mouse.allControls.First(a => a is Vector2Control && a.name == "position") as Vector2Control;
+                    var position = positionControl.ReadValueFromEvent(e);
+
+                    var buttonsClicked = mouse.allControls.FirstOrDefault(a =>
+                        a is ButtonControl abc && abc.ReadValueFromEvent(e) > 0.1f
+                    ) != null;
+                    RGDebug.LogDebug("Mouse event at: " + position.x + "," + position.y + "  buttonsClicked: " + buttonsClicked);
+                    // need to use the static accessor here as this anonymous function's parent gameObject instance could get destroyed
+                    UnityEngine.Object.FindObjectOfType<VirtualMouseCursor>()?.SetPosition(position, buttonsClicked);
+                });
+            }
+
+            return mouse;
+        }
+
+        public static void SendMouseEvent(int replaySegment, MouseInputActionData mouseInput, Dictionary<int, TransformStatus> priorUiTransforms, Dictionary<int, TransformStatus> priorGameObjectTransforms, Dictionary<int, TransformStatus> uiTransforms, Dictionary<int, TransformStatus> gameObjectTransforms)
+        {
+            var clickObjectResult = FindBestClickObject(Camera.main, mouseInput, priorUiTransforms, priorGameObjectTransforms,uiTransforms, gameObjectTransforms);
+
+            var bestObject = clickObjectResult.Item1;
+            var normalizedPosition = clickObjectResult.Item3;
+
+            // non-world space object found, make sure we hit the object
+            if (bestObject != null && bestObject.screenSpaceBounds.HasValue && !clickObjectResult.Item2)
+            {
+                var clickBounds = bestObject.screenSpaceBounds.Value;
+
+                var possibleObjects = clickObjectResult.Item4;
+                foreach (var objectToCheck in possibleObjects)
+                {
+                    if (objectToCheck.screenSpaceBounds.HasValue)
+                    {
+                        var ssb = objectToCheck.screenSpaceBounds.Value;
+                        if (clickBounds.Intersects(ssb))
+                        {
+                            // max of the mins; and min of the maxes
+                            clickBounds.SetMinMax(
+                                Vector3.Max(clickBounds.min, ssb.min),
+                                Vector3.Min(clickBounds.max, ssb.max)
+                            );
+                        }
+                    }
+                }
+
+                // make sure our click is on that object
+                if (!clickBounds.Contains(normalizedPosition))
+                {
+                    var old = normalizedPosition;
+                    // use the center of these bounds as our best point to click
+                    normalizedPosition = new Vector2((int)clickBounds.center.x, (int)clickBounds.center.y);
+
+                    RGDebug.LogInfo($"({replaySegment}) Adjusting click location to ensure hit on object path: " + bestObject.Path + " oldPosition: (" + old.x + "," + old.y + "), newPosition: (" + normalizedPosition.x + "," + normalizedPosition.y + ")");
+                }
+            }
+
+            var mouse = GetMouse();
+
+            using (DeltaStateEvent.From(mouse, out var eventPtr))
+            {
+                eventPtr.time = InputState.currentTime;
+
+                var mouseControls = mouse.allControls;
+                var mouseEventString = "";
+
+                // 1f == true == clicked state
+                // 0f == false == un-clicked state
+                foreach (var mouseControl in mouseControls)
+                {
+                    switch (mouseControl.name)
+                    {
+                        case "delta":
+                            if (_lastMousePosition != null)
+                            {
+                                var delta = normalizedPosition - _lastMousePosition.Value;
+                                if (RGDebug.IsDebugEnabled)
+                                {
+                                    mouseEventString += $"delta: {delta.x},{delta.y}  ";
+                                }
+
+                                ((Vector2Control)mouseControl).WriteValueIntoEvent(delta, eventPtr);
+                            }
+                            break;
+                        case "position":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                mouseEventString += $"position: {normalizedPosition.x},{normalizedPosition.y}  ";
+                            }
+
+                            ((Vector2Control)mouseControl).WriteValueIntoEvent(normalizedPosition, eventPtr);
+                            break;
+                        case "scroll":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.scroll.x < -0.1f || mouseInput.scroll.x > 0.1f || mouseInput.scroll.y < -0.1f || mouseInput.scroll.y > 0.1f)
+                                {
+                                    mouseEventString += $"scroll: {mouseInput.scroll.x},{mouseInput.scroll.y}  ";
+                                }
+                            }
+
+                            ((DeltaControl)mouseControl).WriteValueIntoEvent(mouseInput.scroll, eventPtr);
+                            break;
+                        case "leftButton":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.leftButton)
+                                {
+                                    mouseEventString += $"leftButton  ";
+                                }
+                            }
+
+                            ((ButtonControl)mouseControl).WriteValueIntoEvent(mouseInput.leftButton ? 1f : 0f, eventPtr);
+                            break;
+                        case "middleButton":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.middleButton)
+                                {
+                                    mouseEventString += $"middleButton  ";
+                                }
+                            }
+
+                            ((ButtonControl)mouseControl).WriteValueIntoEvent(mouseInput.middleButton ? 1f : 0f, eventPtr);
+                            break;
+                        case "rightButton":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.rightButton)
+                                {
+                                    mouseEventString += $"rightButton  ";
+                                }
+                            }
+
+                            ((ButtonControl)mouseControl).WriteValueIntoEvent(mouseInput.rightButton ? 1f : 0f, eventPtr);
+                            break;
+                        case "forwardButton":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.forwardButton)
+                                {
+                                    mouseEventString += $"forwardButton  ";
+                                }
+                            }
+
+                            ((ButtonControl)mouseControl).WriteValueIntoEvent(mouseInput.forwardButton ? 1f : 0f, eventPtr);
+                            break;
+                        case "backButton":
+                            if (RGDebug.IsDebugEnabled)
+                            {
+                                if (mouseInput.backButton)
+                                {
+                                    mouseEventString += $"backButton  ";
+                                }
+                            }
+
+                            ((ButtonControl)mouseControl).WriteValueIntoEvent(mouseInput.backButton ? 1f : 0f, eventPtr);
+                            break;
+                    }
+                }
+
+                #if ENABLE_LEGACY_INPUT_MANAGER
+                {
+                    Vector2 delta = _lastMousePosition.HasValue ? (normalizedPosition - _lastMousePosition.Value) : Vector2.zero;
+                    SendMouseEventLegacy(position: normalizedPosition, delta: delta, scroll: mouseInput.scroll,
+                        leftButton: mouseInput.leftButton, middleButton: mouseInput.middleButton, rightButton: mouseInput.rightButton,
+                        forwardButton: mouseInput.forwardButton, backButton: mouseInput.backButton);
+                }
+                #endif
+
+                _lastMousePosition = normalizedPosition;
+
+                if (RGDebug.IsDebugEnabled)
+                {
+                    RGDebug.LogDebug($"({replaySegment}) Sending Mouse Event [{mouseInput.Replay_StartTime}] - {mouseEventString}");
+                }
+
+                InputSystem.QueueEvent(eventPtr);
+            }
+        }
+
+#if ENABLE_LEGACY_INPUT_MANAGER
+        private static void SendMouseEventLegacy(Vector2 position, Vector2 delta, Vector2 scroll,
+            bool leftButton, bool middleButton, bool rightButton, bool forwardButton, bool backButton)
+        {
+            if (RGLegacyInputWrapper.IsPassthrough)
+            {
+                return;
+            }
+
+            RGLegacyInputWrapper.SimulateMouseMovement(new Vector3(position.x, position.y, 0.0f), new Vector3(delta.x, delta.y, 0.0f));
+            RGLegacyInputWrapper.SimulateMouseScrollWheel(scroll);
+
+            if (leftButton)
+                RGLegacyInputWrapper.SimulateKeyPress(KeyCode.Mouse0);
+            else
+                RGLegacyInputWrapper.SimulateKeyRelease(KeyCode.Mouse0);
+
+            if (middleButton)
+                RGLegacyInputWrapper.SimulateKeyPress(KeyCode.Mouse2);
+            else
+                RGLegacyInputWrapper.SimulateKeyRelease(KeyCode.Mouse2);
+
+            if (rightButton)
+                RGLegacyInputWrapper.SimulateKeyPress(KeyCode.Mouse1);
+            else
+                RGLegacyInputWrapper.SimulateKeyRelease(KeyCode.Mouse1);
+
+            if (forwardButton)
+                RGLegacyInputWrapper.SimulateKeyPress(KeyCode.Mouse3);
+            else
+                RGLegacyInputWrapper.SimulateKeyRelease(KeyCode.Mouse3);
+
+            if (backButton)
+                RGLegacyInputWrapper.SimulateKeyPress(KeyCode.Mouse4);
+            else
+                RGLegacyInputWrapper.SimulateKeyRelease(KeyCode.Mouse4);
+        }
+#endif
+
+        // Finds the best object to adjust our click position to for a given mouse input
+        // Uses the exact path for UI clicks, but the normalized path for world space clicks
+        // Returns (the object, whether it was world space, the suggested mouse position)
+        private static (TransformStatus, bool, Vector2, IEnumerable<TransformStatus>) FindBestClickObject(Camera mainCamera, MouseInputActionData mouseInput, Dictionary<int, TransformStatus> priorUiTransforms, Dictionary<int, TransformStatus> priorGameObjectTransforms, Dictionary<int, TransformStatus> uiTransforms, Dictionary<int, TransformStatus> gameObjectTransforms)
+        {
+
+            // Mouse is hard... we can't use the raw position, we need to use the position relative to the current resolution
+            // but.. it gets tougher than that.  Some UI elements scale differently with resolution (only horizontal, only vertical, preserve aspect, expand, etc)
+            // so we have to take the bounding space of the original object(s) clicked on into consideration
+            var normalizedPosition = mouseInput.NormalizedPosition;
+
+            // note that if this mouse input wasn't a click, there will be no possible objects
+            if (mouseInput.clickedObjectNormalizedPaths == null || mouseInput.clickedObjectNormalizedPaths.Length == 0)
+            {
+                // bail out early, no click
+                return (null, false, normalizedPosition, uiTransforms.Values.Concat(gameObjectTransforms.Values));
+            }
+
+            var theNp = new Vector3(normalizedPosition.x, normalizedPosition.y, 0f);
+
+            // find possible objects prioritizing the currentState, falling back to the priorState
+
+            // copy so we can remove
+            var mousePaths = mouseInput.clickedObjectNormalizedPaths;
+            var pathsToFind = mousePaths.ToList();
+
+            var possibleObjects = new List<TransformStatus>();
+
+            foreach (var os in uiTransforms)
+            {
+                var val = os.Value;
+                if (StateRecorderUtils.OptimizedContainsStringInArray(mousePaths, val.Path))
+                {
+                    possibleObjects.Add(val);
+                    StateRecorderUtils.OptimizedRemoveStringFromList(pathsToFind, val.Path);
+                }
+            }
+
+            foreach (var os in gameObjectTransforms)
+            {
+                var val = os.Value;
+                if (StateRecorderUtils.OptimizedContainsStringInArray(mousePaths, val.NormalizedPath))
+                {
+                    possibleObjects.Add(val);
+                    StateRecorderUtils.OptimizedRemoveStringFromList(pathsToFind, val.NormalizedPath);
+                }
+            }
+
+            // still have some objects we didnt' find in the current state, check previous state
+            // this is used primarly for mouse up event processing
+            if (pathsToFind.Count > 0)
+            {
+                foreach (var os in priorUiTransforms)
+                {
+                    var val = os.Value;
+                    if (StateRecorderUtils.OptimizedContainsStringInArray(mousePaths, val.Path))
+                    {
+                        possibleObjects.Add(val);
+                    }
+                }
+
+                foreach (var os in priorGameObjectTransforms)
+                {
+                    var val = os.Value;
+                    if (StateRecorderUtils.OptimizedContainsStringInArray(mousePaths, val.NormalizedPath))
+                    {
+                        possibleObjects.Add(val);
+                    }
+                }
+            }
+
+            var cos = possibleObjects.Where(a=>a.screenSpaceBounds.HasValue).OrderBy(a =>
+            {
+                var closestPointInA = a.screenSpaceBounds.Value.ClosestPoint(theNp);
+                return (theNp - closestPointInA).sqrMagnitude;
+            });
+            possibleObjects = cos.Distinct(_recordedGameObjectStatePathEqualityComparer)
+                .ToList(); // select only the first entry of each path for ui elements; uses ToList due to multiple iterations of this structure later in the code to avoid multiple enumeration
+
+            var screenWidth = Screen.width;
+            var screenHeight = Screen.height;
+            float smallestSize = screenWidth * screenHeight;
+
+            TransformStatus bestObject = null;
+            var worldSpaceObject = false;
+
+            var possibleObjectsCount = possibleObjects.Count;
+            for (var j = 0; j < possibleObjectsCount; j++)
+            {
+                var objectToCheck = possibleObjects[j];
+                var ssb = objectToCheck.screenSpaceBounds.Value;
+                var size = ssb.size.x * ssb.size.y;
+
+                if (objectToCheck.worldSpaceBounds != null)
+                {
+                    if (bestObject != null && bestObject.worldSpaceBounds == null)
+                    {
+                        //do nothing, prefer ui elements
+                    }
+                    else
+                    {
+                        if (mouseInput.worldPosition != null)
+                        {
+                            // use the world space click location closest to the actual object location
+                            var mouseWorldPosition = mouseInput.worldPosition.Value;
+                            // uses the collider bounds on that object as colliders are what would drive world space objects' click detection in scripts / etc
+                            var colliders = objectToCheck.Transform.GetComponentsInChildren<Collider>();
+                            if (colliders.FirstOrDefault(a => a.bounds.Contains(mouseWorldPosition)) != null)
+                            {
+                                var screenPoint = mainCamera.WorldToScreenPoint(mouseWorldPosition);
+                                if (screenPoint.x < 0 || screenPoint.x > screenWidth || screenPoint.y < 0 || screenPoint.y > screenHeight)
+                                {
+                                    RGDebug.LogWarning($"Attempted to click at worldPosition: [{mouseWorldPosition.x},{mouseWorldPosition.y},{mouseWorldPosition.z}], which is off screen at position: [{screenPoint.x},{screenPoint.y}]");
+                                }
+                                else
+                                {
+                                    bestObject = objectToCheck;
+                                    worldSpaceObject = true;
+                                    var old = normalizedPosition;
+                                    // we hit one of our world objects, set the normalized position and stop looping
+                                    normalizedPosition = new Vector2Int((int)screenPoint.x, (int)screenPoint.y);
+                                    RGDebug.LogInfo($"Adjusting world click location to ensure hit on object: " + objectToCheck.Path + " oldPosition: (" + old.x + "," + old.y + "), newPosition: (" + normalizedPosition.x + "," + normalizedPosition.y + ")");
+                                    break; // end the for
+                                }
+                            }
+                            else
+                            {
+                                // didn't hit a collider on this object, we fall back to the renderer bounds bestObject evaluation method, similar to ui elements
+
+                                // compare elements bounds for best match
+                                // give some threshold variance here for floating point math on sizes
+                                // if 2 objects are very similarly sized, we want the one closest, not picking
+                                // one based on some floating point rounding error
+                                if (size * 1.02f < smallestSize)
+                                {
+                                    bestObject = objectToCheck;
+                                    smallestSize = size;
+                                }
+                            }
+                        }
+                        else
+                        {
+                            // mouse input didn't capture a world position, so we weren't clicking on a world space object.. ignore anything not UI related from consideration
+                            if (objectToCheck.worldSpaceBounds == null)
+                            {
+                                // compare elements bounds for best match
+                                // give some threshold variance here for floating point math on sizes
+                                // if 2 objects are very similarly sized, we want the one closest, not picking
+                                // one based on some floating point rounding error
+                                if (size * 1.02f < smallestSize)
+                                {
+                                    bestObject = objectToCheck;
+                                    smallestSize = size;
+                                }
+                            }
+                        }
+                    }
+                }
+                else // objectToCheck.worldSpaceBounds == null
+                {
+                    if (bestObject != null && bestObject.worldSpaceBounds == null)
+                    {
+                        // compare ui elements for best match
+                        // give some threshold variance here for floating point math on sizes
+                        // if 2 objects are very similarly sized, we want the one closest, not picking
+                        // one based on some floating point rounding error
+                        if (size * 1.02f < smallestSize)
+                        {
+                            bestObject = objectToCheck;
+                            smallestSize = size;
+                        }
+                    }
+                    else //bestObject.worldSpaceBounds != null
+                    {
+                        // prefer UI elements when overlaps occur with game objects
+                        bestObject = objectToCheck;
+                        smallestSize = size;
+                    }
+                }
+            }
+
+            return (bestObject, worldSpaceObject, normalizedPosition, possibleObjects);
+        }
+    }
+}
