+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/MouseEventSender.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3450e6f9986c45a5828d152d3a3a6813
+timeCreated: 1718207030
