+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/RecordedGameObjectStatePathEqualityComparer.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9b3c94dc1dbe4e1392cd104980786fd6
+timeCreated: 1718207427
